feat: remove dependencies
