+++ package.json
@@ -1,8 +1,6 @@
 {
   "author": "Marat Zimnurov <zimit@mail.ru>",
-  "dependencies": {
-    "node-crud-kit": "^1.4.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/plugin-syntax-dynamic-import": "^7.0.0",
@@ -39,7 +37,7 @@
     "css-loader": "^4.3.0",
     "cypress": "^5.3.0",
     "dotenv": "^8.2.0",
-    "eslint": "^7.10.0",
+    "eslint": "^7.11.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-import-resolver-typescript": "^2.3.0",
     "eslint-plugin-import": "^2.22.1",
@@ -47,13 +45,14 @@
     "express": "^4.17.1",
     "express-list-endpoints": "^5.0.0",
     "fontello-cli": "^0.5.0",
+    "helmet": "^4.1.1",
     "husky": "^4.3.0",
     "lint-staged": "^10.4.0",
     "npm-run-all": "^4.1.5",
     "open": "^7.3.0",
     "prettier": "^2.1.2",
     "prettier-plugin-svelte": "^1.4.0",
-    "rollup": "^2.28.2",
+    "rollup": "^2.29.0",
     "rollup-plugin-scss": "^2.6.1",
     "rollup-plugin-svelte": "^5.0.1",
     "rollup-plugin-terser": "^7.0.2",
@@ -66,7 +65,7 @@
     "svelte": "^3.29.0",
     "svelte-check": "^0.1.58",
     "svelte-loader": "^2.13.6",
-    "svelte-preprocess": "^4.4.2",
+    "svelte-preprocess": "^4.5.1",
     "ts-loader": "^8.0.4",
     "typescript": "^3.9.7",
     "webpack": "^4.44.2"
